Clean up signed/unsigned mismatch.

+++ client/Game2/RoomList.h
@@ -1,7 +1,7 @@
 
 // RoomList.h
 //
-// Copyright (c) 2009, 2014 Michael Imamura.
+// Copyright (c) 2009, 2014, 2015 Michael Imamura.
 //
 // Licensed under GrokkSoft HoverRace SourceCode License v1.0(the "License");
 // you may not use this file except in compliance with the License.
@@ -91,7 +91,7 @@ class RoomList
 		Server *selectedRoom;
 		banners_t banners;
 		Banner *curBanner;
-		int curBannerIdx;
+		size_t curBannerIdx;
 };
 typedef std::shared_ptr<RoomList> RoomListPtr;
